refactor: 계층에 맞게 폴더 구조 변경
